mlp_trainer/requirements.txt
tensorflow-gpu==2.0.0
numpy>=1.16.4,<2.0
pandas==0.24.0,<1.0
google-cloud-bigquery-storage[pandas,fastavro]==0.7.0
cloudml-hypertune==0.1.0.dev5
talos==0.6.3
google-cloud-bigquery==1.22.0
scipy==1.2.0
ipython==7.9.0
google-cloud-storage==1.23.0
python-snappy==0.5.4
fastavro==0.22.7